move files into the old export dir
