+++ atd-etl/socrata_export/.dockerignore
@@ -1,2 +1,3 @@
 env
 *.env
+**.pyc
