+++ atd-etl/socrata_export/env_template
@@ -1,7 +1,3 @@
-SOCRATA_APP_TOKEN=
-SOCRATA_KEY_ID=
-SOCRATA_KEY_SECRET=
-SOCRATA_DATASET_CRASHES=
-SOCRATA_DATASET_PERSONS=
-HASURA_ADMIN_KEY=
-HASURA_ENDPOINT=
+ENV=dev
+HASURA_GRAPHQL_ENDPOINT="http://localhost:8084/v1/graphql"
+HASURA_GRAPQL_ADMIN_SECRET=hasurapassword
