+++ atd-etl/socrata_export/requirements.txt
@@ -1,1 +1,2 @@
+requests==2.*
 sodapy==2.*
