+++ atd-etl/socrata_export/socrata_export.py
@@ -1,121 +1,93 @@
 #!/usr/bin/env python
-"""
-Socrata - Exporter
-Author: Austin Transportation Department, Data & Technology Services
-
-Description: The purpose of this script is to gather data from Hasura
-and export it to the Socrata database.
-
-The application requires the requests and sodapy libraries:
-    https://pypi.org/project/requests/
-    https://pypi.org/project/sodapy/
-"""
 import os
-import time
-from string import Template
-from sodapy import Socrata
-from process.config import ATD_ETL_CONFIG
-from process.helpers_socrata import *
-from process.socrata_queries import *
 
-print("Socrata - Exporter:  Started.")
+from utils.graphql import make_hasura_request
+from utils.logging import init_logger
+from utils.queries import QUERIES
+from utils.socrata import get_socrata_client
 
-# Setup connection to Socrata
-client = Socrata(
-    "datahub.austintexas.gov",
-    ATD_ETL_CONFIG["SOCRATA_APP_TOKEN"],
-    username=ATD_ETL_CONFIG["SOCRATA_KEY_ID"],
-    password=ATD_ETL_CONFIG["SOCRATA_KEY_SECRET"],
-    timeout=20,
-)
+SOCRATA_DATASET_CRASHES = os.getenv("SOCRATA_DATASET_CRASHES")  # staging: 3aut-fhzp
+SOCRATA_DATASET_PEOPLE = os.getenv("SOCRATA_DATASET_PEOPLE")  # staging: v3x4-fjgm
 
-# Define tables to query from Hasura and publish to Socrata
-query_configs = [
+# number of records to download from Hasura and upload to Socrata
+RECORD_BATCH_SIZE = 100
+# number of times to try a single socrata upload
+MAX_UPLOAD_RETRIES = 3
+
+datasets = [
     {
-        "table": "crash",
-        "template": crashes_query_template,
-        "formatter": format_crash_data,
-        "formatter_config": {
-            "tables": ["atd_txdot_crashes"],
-            "columns_to_rename": {
-                "veh_body_styl_desc": "unit_desc",
-                "veh_unit_desc_desc": "unit_mode",
-                "latitude_primary": "latitude",
-                "longitude_primary": "longitude",
-                "atd_fatality_count": "death_cnt",
-            },
-        },
-        "dataset_uid": ATD_ETL_CONFIG["SOCRATA_DATASET_CRASHES"],
+        "name": "crashes",
+        "typename": "socrata_export_crashes_view",
+        "dataset_id": SOCRATA_DATASET_CRASHES,
+        "query": QUERIES["crashes"],
     },
     {
-        "table": "person",
-        "template": people_query_template,
-        "formatter": format_person_data,
-        "formatter_config": {
-            "tables": ["atd_txdot_person", "atd_txdot_primaryperson"],
-            "columns_to_rename": {"primaryperson_id": "person_id"},
-            "prefixes": {
-                "person_id": "P",
-                "primaryperson_id": "PP",
-            },
-        },
-        "dataset_uid": ATD_ETL_CONFIG["SOCRATA_DATASET_PERSONS"],
+        "name": "people",
+        "typename": "socrata_export_people_view",
+        "dataset_id": SOCRATA_DATASET_PEOPLE,
+        "query": QUERIES["people"],
     },
 ]
 
-# Start timer
-start = time.time()
+# staging
+# https://datahub.austintexas.gov/resource/3aut-fhzp.json?$limit=10
+# https://datahub.austintexas.gov/resource/v3x4-fjgm.json?$limit=10
 
-# For each config, get records from Hasura and upsert to Socrata until res is []
-for config in query_configs:
-    print(f'Starting {config["table"]} table...')
-    is_truncate_complete = False
-    records = None
-    offset = 0
-    limit = 1000
-    total_records = 0
 
-    # Query records from Hasura and upsert to Socrata
-    while records != []:
-        # Create query, increment offset, and query DB
-        query = config["template"].substitute(
-            limit=limit,
-            offset=offset,
-            date_limit=two_weeks_ago(),
-            initial_date_limit=ten_years_ago(),
-        )
-        offset += limit
-        data = run_hasura_query(query)
+def main():
+    socrata_client = get_socrata_client()
 
-        if "data" not in data:
-            print("query: ", query)
-            print("data: ", data)
-            raise RuntimeError("There was a problem gathering data.")
+    for dataset in datasets:
+        logger.info(f"Processing {dataset['name']}")
+        offset = 0
+        records_processed = 0
+        while True:
+            variables = {"limit": RECORD_BATCH_SIZE, "offset": offset}
 
-        # Format records
-        records = config["formatter"](data, config["formatter_config"])
+            logger.info(f"Fetching up to {RECORD_BATCH_SIZE} records - offset={offset}")
+            records = make_hasura_request(query=dataset["query"], variables=variables)[
+                dataset["typename"]
+            ]
 
-        # truncate the existing data set once we have records
-        if records and not is_truncate_complete:
-            print(f'Truncating {config["table"]} table...')
-            client.replace(config["dataset_uid"], [])
-            is_truncate_complete = True
+            if offset == 0:
+                if not records:
+                    raise Exception(
+                        "No records returned from Hasura. This should never happen"
+                    )
+                # we have some records, so it's safe to truncate the Socrata dataset
+                logger.info(f"Truncating dataset {dataset['dataset_id']}")
+                # socrata_client.replace(dataset["dastaset_id"], [])
 
-        # Upsert records to Socrata
-        client.upsert(config["dataset_uid"], records)
-        total_records += len(records)
+            logger.info(f"Upserting {len(records)} records")
 
-        if len(records) == 0:
-            print(f'{total_records} {config["table"]} records upserted.')
-            print(f'Completed {config["table"]} table.')
-        elif total_records != 0:
-            print(f"{total_records} records upserted")
+            upload_error_count = 0
+            while True:
+                # try to upload to socrata until max retries is exceeded - then raise an erro
+                try:
+                    # socrata_client.upsert(dataset["dataset_id"], records)
+                    print("hey")
+                except Exception as e:
+                    upload_error_count += 1
+                    if upload_error_count < MAX_UPLOAD_RETRIES:
+                        logger.error(f"Error uploading to Socrata: {e}. Retrying...")
+                        continue
+                    else:
+                        logger.error(f"Max Socrata upload retries exceeded")
+                        raise e
+                # success - break loop
+                break
 
-# Terminate Socrata connection
-client.close()
+            records_processed += len(records)
 
-# Stop timer and print duration
-end = time.time()
-hours, rem = divmod(end - start, 3600)
-minutes, seconds = divmod(rem, 60)
-print("Finished in: {:0>2}:{:0>2}:{:05.2f}".format(int(hours), int(minutes), seconds))
+            if len(records) < RECORD_BATCH_SIZE:
+                logger.info(f"{records_processed} {dataset['name']} records processed")
+                break
+
+            offset += RECORD_BATCH_SIZE
+
+    return
+
+
+if __name__ == "__main__":
+    logger = init_logger()
+    main()
