--- atd-etl/socrata_export_v2/.dockerignore
@@ -1,3 +1,0 @@
-env
-*.env
-**.pyc
