--- atd-etl/socrata_export_v2/Dockerfile
@@ -1,8 +1,0 @@
-FROM python:3.11-slim
-
-WORKDIR /app
-COPY . /app
-
-RUN apt-get -y update
-RUN chmod -R 755 /app/*
-RUN pip install -r requirements.txt
