--- atd-etl/socrata_export_v2/README.md
@@ -1,1 +1,0 @@
-# Socrata Open Data Portal Export
