--- atd-etl/socrata_export_v2/docker-compose.yaml
@@ -1,9 +1,0 @@
-services:
-  socrata_export:
-    build: .
-    volumes:
-      - .:/app
-    env_file:
-      - .env
-    network_mode: host
-    entrypoint: /bin/bash
