--- atd-etl/socrata_export_v2/requirements.txt
@@ -1,2 +1,0 @@
-requests==2.*
-sodapy==2.*
